add signout button to modal if signed in but not a subscriber
