+++ app/javascript/app/Header/Profile/PatreonModal/BecomeAPatronButton/BecomeAPatronButton.svelte
@@ -1,0 +1,28 @@
+<script>
+	import PatronButton from '../../../../../assets/become_a_patron_button@2x.png'
+</script>
+
+<a
+  class='become-patron-button'
+  href='https://www.patreon.com/catalyticsound'
+  target="_blank" rel="noopener noreferrer">
+  <img
+    class='become-patron-img'
+    src={PatronButton}
+    alt='become a patron' />
+</a>
+
+<style>
+  .become-patron-button {
+		width: 200px;
+		height: 50px;
+		border: none;
+		margin: 8px;
+
+	}
+
+	.become-patron-img {
+		width: 100%;
+		height: 100%;
+	}
+</style>
